add norm and change float to double

+++ src/CLI/CLI.java
@@ -7,13 +7,13 @@ import utils.Node;
 
 public class CLI {
 
-	private List<Float> c;
+	private List<Double> c;
 	private List<Integer> l;
 	private List<Integer> i;
 	private int nbNodes;
 	
 	public CLI(int nbNodes){
-		this.c = new ArrayList<Float>();
+		this.c = new ArrayList<Double>();
 		this.l = new ArrayList<Integer>();
 		this.i = new ArrayList<Integer>();
 		this.nbNodes = nbNodes;
@@ -68,7 +68,7 @@ public class CLI {
 
 	public void addC(List<Node> listNode){
 		for(int j = 0 ; j < listNode.size(); j++){
-			this.c.add((float) 1 / listNode.size());
+			this.c.add((double) 1 / listNode.size());
 		}
 	}
 	
@@ -140,11 +140,11 @@ public class CLI {
 		this.nbNodes = nbNodes;
 	}
 
-	public List<Float> getC() {
+	public List<Double> getC() {
 		return c;
 	}
 
-	public void setC(List<Float> c) {
+	public void setC(List<Double> c) {
 		this.c = c;
 	}
